chore(db): move db migrations into own file
